+++ src/main/java/com/drakmyth/minecraft/manufactory/ModEventSubscriber.java
@@ -14,9 +14,6 @@ import com.drakmyth.minecraft.manufactory.datagen.ModLanguageProvider;
 import com.drakmyth.minecraft.manufactory.datagen.ModLootTableProvider;
 import com.drakmyth.minecraft.manufactory.datagen.ModRecipeProvider;
 import com.drakmyth.minecraft.manufactory.datagen.ModTagsProvider;
-import com.drakmyth.minecraft.manufactory.gui.PowerProgressScreen;
-import com.drakmyth.minecraft.manufactory.gui.ScreenTextures;
-import com.drakmyth.minecraft.manufactory.gui.SimpleScreen;
 import com.drakmyth.minecraft.manufactory.init.ModBlocks;
 import com.drakmyth.minecraft.manufactory.init.ModMenuTypes;
 import com.drakmyth.minecraft.manufactory.menus.BallMillMenu;
@@ -24,6 +21,9 @@ import com.drakmyth.minecraft.manufactory.menus.BallMillUpgradeMenu;
 import com.drakmyth.minecraft.manufactory.menus.GrinderMenu;
 import com.drakmyth.minecraft.manufactory.menus.GrinderUpgradeMenu;
 import com.drakmyth.minecraft.manufactory.menus.RockDrillUpgradeMenu;
+import com.drakmyth.minecraft.manufactory.menus.screens.PowerProgressScreen;
+import com.drakmyth.minecraft.manufactory.menus.screens.ScreenTextures;
+import com.drakmyth.minecraft.manufactory.menus.screens.SimpleScreen;
 import com.drakmyth.minecraft.manufactory.init.ModBlockEntityTypes;
 
 import org.apache.logging.log4j.LogManager;
