+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/screens/PowerProgressScreen.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
 
-package com.drakmyth.minecraft.manufactory.gui;
+package com.drakmyth.minecraft.manufactory.menus.screens;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
 import com.drakmyth.minecraft.manufactory.menus.IPowerProgressMenu;
