+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/screens/ScreenTextures.java
@@ -2,7 +2,7 @@
  *  SPDX-License-Identifier: LGPL-3.0-only
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
-package com.drakmyth.minecraft.manufactory.gui;
+package com.drakmyth.minecraft.manufactory.menus.screens;
 
 import java.util.Map;
 import static java.util.Map.entry;
